+++ client/src/types/global-cloudflare.d.ts
