+++ client/src/util/lang.ts
@@ -12,7 +12,7 @@ export function compactObject<T>(obj: Record<string, T | null | undefined>): Rec
 
 export type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
 
-export function isObject(value: any): value is object {
+export function isObject(value: any): value is Record<string, unknown> {
   return value !== undefined && value !== null && typeof value === 'object';
 }
